rewrite insert_newline for brevity
This effectively isolates the code that finds the indentation length,
which will later be extracted out for use in reflowing.

This also adds a test of newline insertion with indentation maintenence
when the target line is in the middle of a document, to verify that the
subroutine is properly identifying the line.