reflow around indents
Text that's indented (say, docstrings inside an Impl) can be reflowed
just the same way as something unindented like a commit message. This
addition is transparent and backwards-compatible.

This has less-than-good unicode support, and I can see some breakages on
weird unicode whitespace (maybe the 0-width space counting as a
character) in the future, but I honestly don't know if that's something
anyone's going to run into.