package with Nix
This is for personal use. Do not depend on this. It is ugly.

I'm *definitely* concerned that I needlessly screwed with some
dependencies, and most of the past three commits are at best going to
cause merge conflicts down the line while providing no utility
whatsoever. If I were smart, I'd go through and fix then rebase all
this, but all I actually care about is using this Amp fork in my NixOS
config.